section2/requirements.txt
web3>=6.20.0
python-dotenv>=1.0.1
eth-account>=0.11.2
rich>=13.7.1
